directXtex

+++ main.cpp
@@ -1,9 +1,10 @@
 #include <cstdio>
 #include <cstdlib>
+#include <DirectXTex.h>
 
-int main(int argc,char* argv[])
+size_t main(size_t argc,char* argv[])
 {
-	for (int i = 0; i < argc; i++)
+	for (size_t i = 0; i < argc; i++)
 	{
 		printf(argv[i]);
